requirements.txt
flask
spotipy
flask-wtf
wtforms
python-dotenv
cors
flask-cors